ci: store artifacts after build
# Motivation

To be able to build and release more automatic and
less dependent on the local environment, we need
to store the build artifacts in a place where we
can access them later.

# Changes

Stores some of the build artifacts from the `dist`
directory after a successful build.

# Future work

Use these artifacts to create a release together
with autogenerated docs.

# Quirks

Ideally the artifacts should be built in a
separate job to avoid potential contamination of
the build environment. But because this is a paid
`gorelease` feature we will just have to continue
building all targets in one go.